+++ libtiledbvcf/src/read/reader.cc
@@ -962,7 +962,7 @@ void Reader::prepare_variant_stats() {
 void Reader::read_from_variant_stats(
     uint32_t* pos,
     char* allele,
-    uint64_t* allele_offsets,
+    int32_t* allele_offsets,
     int* ac,
     int* an,
     float_t* af) {
